refactor: modularize SessionHistoryPanel and fix status sorting priority bug
